+++ tools/src/lib.rs
@@ -10,4 +10,3 @@ pub mod sliced_diff;
 pub mod supports;
 pub mod test_pattern;
 pub use exposure::{exposure_test, internal_exposure_test};
-pub mod sdf;